requirements.txt
alembic==1.11.3
Faker==19.3.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.12